pyproject.toml
[project]
name = "kingshot-role-manager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.7.1",
]
